Extend PXE boot retry timeout for RAX hosts
When extending the timeouts for jobs to execute with-in,
we've observed a case where RAX hosts are cutting off at
the time limit of 900 seconds (as being asserted by another
change set). This is both good and bad. We know the timeout
feature works, but the agent was not quite online yet.

As such, we should also auto-extend base retry timeouts
so there is hope for the job to complete.

Change-Id: I8efa3a52188de558a7964d1daafd2225e102e251