Add expose body decorator, status_code argument
@method.body populates a method argument with the contents of the
request JSON. When the request body JSON is a dict, pecan populates
the keyword arguments of the method with the values of the dict. This
is not desirable for our typical usage so the body decorator also
strips these keyword arguments.

A status_code argument is also added to @method.expose to set custom
responses. Empty response status codes are given special handling.

This change also adds unit tests for the @method.expose decorator.

Change-Id: I82d75345e4512c79ad7d211e81ca474f1b45a7ca
Story: 1651346
Task: 10551