Disable the multinode job for queens
While we would prefer it to be working and voting,
we seem to be fighting some sort of issue on multiple
branches where authentication is now failing without
any real version change.

Given the age of this branch, and the real purpose just to
provide basic testing for older backports, we're disabling
the job at this time.

Change-Id: I3fe02bd88bcde567607b0d1992d85d98361516aa