Handle an older agent with agent_token
Well, as the author of the agent token work, I could have
sworn we covered this case, but it seems not and agent commands
could fail.

What was occuring, when the token is optional, we were not backing
down on the agent client and failing even though we could detect
the failure and handle it accordingly.

Change-Id: Ibd7e17fb00747485c8072289fff9b28d4da17c39
Story: 2008002
Task: 40649