Handle an older agent with agent_token
Well, as the author of the agent token work, I could have
sworn we covered this case, but it seems not and agent commands
could fail.

What was occuring, when the token is optional, we were not backing
down on the agent client and failing even though we could detect
the failure and handle it accordingly.

Note: This change had to be slightly re-worked and thus was not a
perfectly clean backport due to changes made earlier in the victoria
development cycle in I62e9086441fa2b164aee42f7489d12aed4076f49.
As a result, the helper class for faults is not present in this
change set and the messages returned from the tests had to be
passed slightly differently using the original FakeResponse class.

Change-Id: Ibd7e17fb00747485c8072289fff9b28d4da17c39
Story: 2008002
Task: 40649
(cherry picked from commit 30d9cb47e62b62d570e1792515e16abf1ac3cd56)
(cherry picked from commit 980bb3f9465746e34b846f1e337ccceaccbf1bc7)