Minor agent version code cleanup
Internal to the conductor, we have long determiend that if no
agent version is submitted by the client, that the client is a
version 3.0.0 agent.

With agent token, we're effectively requiring 6.1.0 or later
to be leveraged, so all agents should be sending versions.

An agent not sending versions is thus unsupported and would no
longer work without agent token support.

Redudnant code has thus been removed. Additionally the conductor
utility is_agent_token_supported has been removed since it is now
redundant.

Change-Id: Id6c8d1df08c3ac7af61ed7d05d274f3099003582