Clarify HPE Edgeline support
Per discussions with stendulker, the Edgeline servers are not
recognized by proliantutils which means they are not handled
properly and the wrong devices can be set/selected for setting
the next boot device. This is largely because proliantutils
thinks Edgeline hardware is 8th generation instead of 10th
generation hardware.

Since the docs explicitly state Proliant line of rackmount
servers, it seems logical to clarify the Edgeline servers
as it is a different line of servers with a different
market focus.

Change-Id: I75062be1799ea5dd906e675d1b53c10d4f23ce3a