CI: Remove the build check for pre-build ramdisks only
In the devstack plugin, we identify RAX hosts and change the settings
slightly to help ensure CI job passage. This is because the machines are
fully emulated as opposed to paravirtualized which causes a performance
impact with our test VMs that we create. Since we can hit issues
while scheduling on RAX in general, it only makes sense to also
swap the job out for centos based IPA jobs that some how happen on
to a RAX node. Since the CI jobs have largely been "de-typed" name
wise, this should be fine.

Change-Id: I63fc5de05c3975433c12ddd9ba21ebb676de2e6c