Reduce grenade node count
Infra's disk/swap availability has been apparently
reduced with the new focal node sets such that we
have ~60GB of disk space and only 1GB of swap.

If we configure more swap, then naturally that means
we take away from available VMs as well.

And as such, we should be able to complete grenade
with only four instances, I hope.

Change-Id: I36f8fc8130ed914e8a2c2a11c9679144d931ad73