CI: Additional minor job changes to improve CI reliability
Automated cleaning makes sense to test and ensure working in
more complex and focused scenarios, but realistically we don't
need to run additional integration testing when it is such a core
component of ironic's behavior.

Also marked the ironic-tempest-ipa-partition-uefi-pxe_ipmitool
job as non-voting. By default it is trying to run with 4GB VMs
which means it was always running the risk of being over comitted
once we introduced the second VM even if it is never started.

Also reduces the size of the UEFI VM RAM as the latest ussuri
ramdisks are now smaller due to the additional cleanup added in
IPA-B. This now matches master branch large VM default.

Also increases general swap size to 4GB so hopefully we can at
least maybe have a semi-funcitonal gate.

Change-Id: I57f36631408f35275eb450e9a28f9c766f4be9b8