Clarify power state logging
Some leftover logging displays a warning that the power state is ERROR
even when it isn't.

This change has a new logging message which includes the current and
requested power state. The logging is info instead of warning because
it is expected that the requested state is different from the power
state.

Change-Id: I4e592da89fa36963a66ebb1adfe6070dd13a6e6e