Log the ports we bind in flat
While trying to troubelshoot the standalone job's rescue
failures, I noticed we didn't log the VIF port ID of the
port we use for rescue. Which makes it far more difficult
to backtrack and try and figure out where things went sideways.

So it seems clear we should log the ports we succeed at binding,
and thus the flat driver now keeps a quick list of ports bound
and will log that as a result.

Change-Id: Id66e5398ee3a1e1630262a229ee6c706b7e62a08