IPMI: Handle vendor set boot device differences
Supermicro machines, when in UEFI mode, have a different
device number, in binary, to represent the hard disk from
other vendors such as Fujitsu which actually has somewhat
similar code in their driver.

This means we need to be somewhat cognizent of the vendor of
the BMC and possibly update the device mapping based upon that
vendor.

This may ultimately fix a number of IPMI related problems, because
there is a reliance upon the text output of ipmitool, which only
reads the bytes retured by the BMC, which may not be reality after
the next reset, espescialy if ipmitool doesn't know of the UEFI
operating difference.

NOTE: This had to be a direct cherry-pick from the original change
as the intermediate branch that we normally would have cherry-picked
from has had other changes made to it, which resulted in an unclean
backport, where as with the original change it was clean.

Change-Id: Ie19db9e0cf1eafdfc9bb46248f4d457337821f94
Story: 2008241
Task: 41085
(cherry picked from commit a7ac9ce8cd53fdebdf0d9aaeaf439da08a0b3ec8)