Simplify injecting network data into an ISO image
Currently we're building a VFAT image with the network data just
to unpack it back on the next step. Just pass the file directly.
This fixes a permission denied problem on Bifrost on Fedora
(at least).

As a nice side effect, the change reduces the amount of IO done
for virtual media quite substantially.

Change-Id: I5499fa42c1d82a1a29099fbbba6f45d440448b72