Sync boot mode when changing the boot device via Redfish
After changing the boot device via Redfish, check that the boot mode being
reported matches what is configured and, if not, set it to the configured
value.  Some BMCs change the boot mode when the device is set via Redfish,
this will ensure the mode is set properly.

Change-Id: Ib077f7f32de029833e6bd936853c382305bce36e
Story: 2008252
Task: 41103