Fixes the issue that instance bond port can't get IP address
The issue is that when a port group doesn't have a mac address assigned by
operators, and during provisioning we unbind/bind tenant port with None which
causes the mac address to be regenerated twice and differs from the originally
one allocated by nova or users which was packed into config drive. The end result
is that, bond port has different mac address configured and can't the IP address
from neutron.

Change-Id: I92ed5d17239216324d6a69e0ed8771fd6948d6ec
Story: 2008300
Task: 41185