Fix disk label to account for UEFI
Previously disk labels would not be populated if not explicitly
set by an API user, which lead to a dangerous possible case,
which sometimes could work, but was ultimately wrong to
setup a UEFI booting machine with a BIOS MBR partition table.

Not all systems support this, but UEFI systems are supposed to
support GPT partition tables.

We now fallback if no explicit override is set and assume GPT
if the machine is set to UEFI mode.

Change-Id: I001d8c6ee3b1d6c466c71ea5179bdbca9bdd692d