Use ddt for existing policy tests
This change aims to replicate the existing policy check tests using
ddt so that they can be represented in a declarative format.

The number of tests are expected to grow significantly with
secure-rbac, so the aim of this change is to explore how best to
represent the tests before the new tests are written.

Change-Id: Id1b60f5525773eadbe1b85797eeefdb70d73ae1d