Duplicate testing for system scoped ACL testing
In order for us to understand if we're starting to break
compatability with scope changes, we need to at least have
some testing, someway, some place to provide insight.

Hopefully, this should provide that insight by running
the same API testing we run with by default.

Once we have fully defined the scoped authenticaiton
use requirements, we can begin to build the new test
matrix, but this gives us a start.

Change-Id: Ie12e2665c97326142b564b23603647ec9d97052c