For Supermicro BMCs set enable when changing boot device
When setting the boot settings on Supermicro BMCs, the
BootSourceOverrideEnabled must be set to the desired
value whenever the BootSourceOverrideTarget is set or
it will revert to the default value (Once).  This is
different than what is currently implemented for other
BMCs in which the BootSourceOverrideEnabled is not set
if it matches the current setting.

This change uses the vendor setting to determine if it's
a Supermicro BMC.

Story: 2008547
Task: 41652

Change-Id: I1b1a6baafd4cc4daa2fbdb82f69ded6253b1fcbf
(cherry picked from commit 4a7d50ce56b88c82e95665ac6df44666d3e647ed)