RBAC System Scope: observer -> reader
One of the default role names in the RBAC model with system and
project scopes is reader. Reader replaces observer, and while this
was not done earlier to the tests in system scope was because it is
better to evolve the tests being able to run individual groups with
the same name as opposed to different names.

Change-Id: I57bab93adaf7e562c4c46febd612e1f27ea50bfa