Allow running RAID cleaning steps with in-band cleaning
In I3b647e39982c0a98abac7b0a7c1c60215d6db4f2, the change
allowed users to set priorities for create and delete config
however it was not set in the right place because the steps live
in the `raid` interface.

This patch introduces a get_clean_steps for the RAID interface
which gets all the steps and the correct overrides.  This has been
tested to work in a local environment.

Change-Id: I5138ec9daf27affc28647dd83c79a867c9be6f3d