Fix webserver_verify_ca config documentation
When a specific configuration options list is created and delineated
out, it needs to be explicitly added so it gets pulled in, either
through each configuration generator file's list options method, or
in the case of default for the list of lists to iterate through.

It seems websever_verify_ca was missing in the docs, and this
results in it being properly generated.

Change-Id: I84709f5093c4b716b4cb21bd31fa05c65f93d8df
Story: 2008732
Task: 42084