Fix configuration generation for ironic doc pages
Ironic's sample configuration page previously did not render any
of the items in the default section, except for those items added
by other libraries. This was because we were trying to use an iterator
instead of a list.

Using an iterator, in theory should have worked, and did work for
normal invocations, but didn't work when it came to sphinx generated
output.

Instead of trying to use itertools to assemble everything, we just
now instead assemble the list and use a list_opts method like some
of the other more complex groups to add values.

Confirmed in local build output that the sphinx generated output
works as expected now.

Change-Id: I7f1cffb2a91728ab632ab0ccaa6acbb7e86fb533